issue #109 InvItems finished (prices, stocks, validations, exrates conversions), AdmExchageRate finished, Menu when index bug fixed

+++ app/Controller/InvItemsController.php
@@ -38,7 +38,9 @@ class InvItemsController extends AppController {
 		$filters = array();
 		$code = '';
 		////////////////////////////START - WHEN SEARCH IS SEND THROUGH POST//////////////////////////////////////
+		
 		if($this->request->is("post")) {
+			
 			$url = array('action'=>'index');
 			$parameters = array();
 			$empty=0;
@@ -48,7 +50,13 @@ class InvItemsController extends AppController {
 				$empty++;
 			}
 			
-			if($empty == 1){
+//			if(isset($this->request->data['InvItem']['stock']) && $this->request->data['InvItem']['stock']){
+				$parameters['stock'] = $this->request->data['InvItem']['stock']; //this case is different (witout validation because is select and always has a value)
+//			}else{
+//				$empty++;
+//			}
+			
+			if($empty == 2){
 				$parameters['search']='empty';
 			}else{
 				$parameters['search']='yes';
@@ -58,6 +66,7 @@ class InvItemsController extends AppController {
 		////////////////////////////END - WHEN SEARCH IS SEND THROUGH POST//////////////////////////////////////
 		
 		////////////////////////////START - SETTING URL FILTERS//////////////////////////////////////
+//		debug($this->passedArgs);
 		if(isset($this->passedArgs['code'])){
 			$filters['InvItem.code LIKE'] = '%'.strtoupper($this->passedArgs['code']).'%';
 			$code = $this->passedArgs['code'];
@@ -82,7 +91,19 @@ class InvItemsController extends AppController {
 		foreach($pagination as $val){
 			$items[$val['InvItem']['id']] = $val['InvItem']['id'];
 		}
-		$stocks = $this->_get_stocks($items);
+		
+		$valueWarehouse = 0;
+		
+		if(isset($this->passedArgs['stock'])){
+			if($this->passedArgs['stock'] == 0){
+				$stocks = $this->_get_stocks($items);
+			}else{
+				$stocks = $this->_get_stocks($items, $this->passedArgs['stock']);
+				$valueWarehouse = $this->passedArgs['stock'];
+			}
+		}else{
+			$stocks = $this->_get_stocks($items);
+		}
 		//the array format is like this:
 		/*
 		array(
@@ -100,9 +121,15 @@ class InvItemsController extends AppController {
 				$pagination[$key]['InvItem']['stock'] = $this->_find_item_stock($stocks, $val['InvItem']['id']);
 				//debug( $this->_find_item_stock($stocks, $val['InvItem']['id']));
 		}
+		
+		$this->loadModel("InvWarehouse");
+		$warehouses = $this->InvWarehouse->find("list");
+		$warehouses[0] = "Todos los almacenes";
 		//debug($pagination);
 		////////////////End - Stocks		
-				
+//				debug($warehouses);
+		$this->set("valueWarehouse", $valueWarehouse);
+		$this->set("warehouses",$warehouses);		
 		$this->set('invItems', $pagination);
 		$this->set('code', $code);
 		////////////////////////END - SETTING PAGINATE AND OTHER VARIABLES TO THE VIEW//////////////////
@@ -173,7 +200,13 @@ class InvItemsController extends AppController {
  * @return void
  */
 	public function add() {
+		
 		//Section where the controls of the page are loaded		
+		$invSuppliers = $this->InvItem->InvItemsSupplier->InvSupplier->find('list', array('order' => 'InvSupplier.name'));
+		if(count($invSuppliers) == 0)
+		{
+			$invSuppliers[""] = '--- Vacio ---';
+		}
 		$invBrands = $this->InvItem->InvBrand->find('list', array('order' => 'InvBrand.name'));
 		if(count($invBrands) == 0)
 		{
@@ -185,32 +218,74 @@ class InvItemsController extends AppController {
 		{
 			$invCategories[""] = '--- Vacio ---';
 		}		
-		$this->set(compact('invBrands', 'invCategories'));	
+		$this->set(compact('invBrands', 'invCategories', 'invSuppliers'));	
 		
 		
 		//Section where information is saved into the database
 		if ($this->request->is('post')) {			
-			$this->InvItem->create();			
-			if ($this->InvItem->save($this->request->data)) {
-				$this->Session->setFlash(
-					__('El Item se guardo satisfactoriamente'),
-					'alert',
-					array(
-						'plugin' => 'TwitterBootstrap',
-						'class' => 'alert-success'
-					)
-				);
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(
-					__('El Item no se pudo guardar, por favor intente de nuevo'),
-					'alert',
-					array(
-						'plugin' => 'TwitterBootstrap',
-						'class' => 'alert-error'
-					)
-				);
+			
+			$data = $this->request->data;
+//			debug($data);
+			
+			$this->loadModel("AdmExchangeRate");
+			$existexRate = $this->AdmExchangeRate->find("first", array(
+				"conditions"=>array("AdmExchangeRate.date <="=>$data["InvPrice"][0]["date"]),
+				"fields"=>array("AdmExchangeRate.value")
+			));
+			$exRate = 0;
+//			debug($existexRate);
+			$error=0;
+			if(count($existexRate) == 0){
+				$this->Session->setFlash('No hay un "Tipo de Cambio" registrado para la fecha elegida','alert',array('plugin' => 'TwitterBootstrap','class' => 'alert-error'));
+				//$this->redirect(array('action' => 'add'));
+				$error++;
+				
+			}else{
+				$exRate = $existexRate["AdmExchangeRate"]["value"];
 			}
+			
+			
+			
+			if($error == 0){
+				$data["InvPrice"][0]["ex_rate"] = $exRate;
+				$data["InvPrice"][0]["inv_price_type_id"] = 1;
+				$data["InvPrice"][0]["description"] = "Precio FOB Inicial";
+				if($data["Aux"]["priceType"] == "BOLIVIANOS"){
+					$data["InvPrice"][0]["price"] = $data["Aux"]["neutralPrice"];
+					$data["InvPrice"][0]["ex_price"] = $data["Aux"]["neutralPrice"] / $exRate;
+				}elseif($data["Aux"]["priceType"] == "DOLARES"){
+					$data["InvPrice"][0]["price"] = $data["Aux"]["neutralPrice"] * $exRate;
+					$data["InvPrice"][0]["ex_price"] = $data["Aux"]["neutralPrice"];
+				}
+				
+				//converting to UPPER for code
+				$data["InvItem"]["code"] = strtoupper($data["InvItem"]["code"]);
+			
+	//			debug($data);
+				//$data["InvItemsSupplier"]
+				$this->InvItem->create();			
+
+				if ($this->InvItem->saveAll($data, array("deep"=>true))) {
+					$this->Session->setFlash(
+						__('El Producto se guardo satisfactoriamente'),
+						'alert',
+						array(
+							'plugin' => 'TwitterBootstrap',
+							'class' => 'alert-success'
+						)
+					);
+					$this->redirect(array('action' => 'add'));
+				} else {
+					$this->Session->setFlash(
+						__('El Producto no se pudo guardar, por favor intente de nuevo'),
+						'alert',
+						array(
+							'plugin' => 'TwitterBootstrap',
+							'class' => 'alert-error'
+						)
+					);
+				}
+			}//end if error
 		}
 		
 		
@@ -283,7 +358,7 @@ class InvItemsController extends AppController {
 		}
 		if ($this->InvItem->delete()) {
 			$this->Session->setFlash(
-				__('El Item fue Eliminado'),
+				__('El Producto fue Eliminado'),
 				'alert',
 				array(
 					'plugin' => 'TwitterBootstrap',
@@ -293,7 +368,7 @@ class InvItemsController extends AppController {
 			$this->redirect(array('action' => 'index'));
 		}
 		$this->Session->setFlash(
-			__('El Item no se pudo Eliminar'),
+			__('El Producto no se pudo Eliminar'),
 			'alert',
 			array(
 				'plugin' => 'TwitterBootstrap',
@@ -308,7 +383,8 @@ class InvItemsController extends AppController {
 	public function ajax_initiate_modal_add_price(){
 		if($this->RequestHandler->isAjax()){
 						
-			$pricesAlreadySaved = $this->request->data['pricesAlreadySaved'];
+//			$pricesAlreadySaved = $this->request->data['pricesAlreadySaved'];
+			
 //			$warehouse = $this->request->data['warehouse']; //if it's warehouse_transfer is OUT
 //			$warehouse2 = $this->request->data['warehouse2'];//if it's warehouse_transfer is IN
 //			$transfer = $this->request->data['transfer'];
@@ -316,6 +392,8 @@ class InvItemsController extends AppController {
 			$invPriceTypes = $this->InvItem->InvPrice->InvPriceType->find('list', array(
 				//'recursive'=>-1,
 				//'fields'=>array('InvItem.id', 'CONCAT(InvItem.code, '-', InvItem.name)')
+				//"order"=>array("InvPriceType.name"),
+				"conditions"=>array("InvPriceType.name"=>array("FOB", "CIF", "VENTA"))
 			));	
 			
 			$this->set(compact('invPriceTypes'));
@@ -330,27 +408,71 @@ class InvItemsController extends AppController {
 			$priceDate = $this->request->data['priceDate'];
 			$priceAmount = $this->request->data['priceAmount'];			
 			$priceDescription = $this->request->data['priceDescription'];			
+			$currencyType = $this->request->data['currencyType'];			
 			
-			$arrayPrice = array('inv_item_id'=>$itemId, 'inv_price_type_id'=>$priceTypeId, 'date'=> $priceDate, 'price'=>$priceAmount, 'description'=>$priceDescription);
-			$data = array('InvPrice'=>$arrayPrice);
+			
+			$this->loadModel("AdmExchangeRate");
+			$existexRate = $this->AdmExchangeRate->find("first", array(
+				"conditions"=>array("AdmExchangeRate.date <="=>$priceDate),
+				"fields"=>array("AdmExchangeRate.value")
+			));
+			
+			
+			
+			
+			
+			if(count($existexRate) > 0){
+				$exRate = $existexRate["AdmExchangeRate"]["value"];
+				
+				if($currencyType == "BOLIVIANOS"){
+					$price = $priceAmount;
+					$exPrice = $priceAmount / $exRate;
 
-			if($this->InvItem->InvPrice->saveAssociated($data)){
-					$priceIdInserted = $this->InvItem->InvPrice->id;
-						echo 'insertado|'.$priceIdInserted;
+				}elseif($currencyType == "DOLARES"){
+					$price = $priceAmount * $exRate;
+					$exPrice = $priceAmount;
 				}
+				
+				
+				$arrayPrice = array('inv_item_id'=>$itemId, 'inv_price_type_id'=>$priceTypeId, 'date'=> $priceDate, 'price'=>$price, 'ex_price'=>$exPrice, 'ex_rate'=>$exRate, 'description'=>$priceDescription);
+				$data = array('InvPrice'=>$arrayPrice);
+
+				if($this->InvItem->InvPrice->saveAssociated($data)){
+//						$priceIdInserted = $this->InvItem->InvPrice->id;
+//							echo 'insertado|'.$priceIdInserted;
+						echo "success";
+					}
+			}else{
+				echo "noExRate";
+			}
+			
+			
 		}
 	}
 	
+
 	
 	public function ajax_delete_price(){
 		if($this->RequestHandler->isAjax()){			
 			
 			$priceId = $this->request->data['priceId'];			
+			$itemId = $this->request->data['itemId'];			
+			$priceTypeId = $this->request->data['priceTypeId'];
 			
-			$arrayPrice = array('inv_price_id'=>$priceId);
-			$data = array('InvPrice'=>$arrayPrice);
+			$exists = $this->InvItem->InvPrice->find("count", array("conditions"=>array("InvPrice.inv_item_id"=>$itemId, "InvPrice.inv_price_type_id"=>$priceTypeId)));
 			
-			$this->InvItem->InvPrice->deleteAll(array('InvPrice.id' => $arrayPrice));
+			//debug()
+			
+			if($exists > 1){
+				$arrayPrice = array('inv_price_id'=>$priceId);
+				//$data = array('InvPrice'=>$arrayPrice);
+				$this->InvItem->InvPrice->deleteAll(array('InvPrice.id' => $arrayPrice));
+				echo "success";
+			}else{
+				echo "mustExistOne";
+			}
+			
+			
 
 		}
 	}
@@ -360,7 +482,7 @@ class InvItemsController extends AppController {
 			
 			$itemId = $this->request->data['itemId'];
 			$itemSupplier = $this->request->data['itemSupplier'];
-			$itemCode = $this->request->data['itemCode'];
+			$itemCode = strtoupper($this->request->data['itemCode']);
 			$itemBrand = $this->request->data['itemBrand'];
 			$itemCategory = $this->request->data['itemCategory'];
 			$itemName = $this->request->data['itemName'];
@@ -408,7 +530,7 @@ class InvItemsController extends AppController {
 	private function _get_prices($idPrice){
 		$invPrices = $this->InvItem->InvPrice->find('all',array(
 			'conditions' => array('InvPrice.inv_item_id' => $idPrice),
-			'fields' => array('InvPrice.id','InvPriceType.name','InvPrice.date','InvPrice.price','InvPrice.description'),
+			'fields' => array('InvPrice.id','InvPriceType.name','InvPrice.date','InvPrice.price', 'InvPrice.ex_price','InvPrice.description', 'InvPriceType.id'),
 			'order' => array('InvPrice.date' => 'desc'),
 		));
 		
@@ -418,8 +540,10 @@ class InvItemsController extends AppController {
 				'itemId' => $idPrice,
 				'priceId' => $value['InvPrice']['id'],
 				'priceType' => $value['InvPriceType']['name'],
+				'priceTypeId' => $value['InvPriceType']['id'],
 				'date' => $value['InvPrice']['date'],
 				'price' => $value['InvPrice']['price'],
+				'ex_price' => $value['InvPrice']['ex_price'],
 				'description' => $value['InvPrice']['description']
 			);
 		}
